Updated the paths to the libs
The error now is that it cannot open libopenblas.lib. Maybe that is
because it didn't know where to look. Maybe that's because it's a linux
library and windows doesn't know what to do. I'm going to try to
resolve the former and if that doesn't work then the latter.