Added ncurses as a build
This one was complicated to build locally. Because of that I'm taking a different approach and including the build dependencies in the repo. The dependencies hosted on PHP's site are out of date, and do not build.

That said it might be a good idea to have all of the build deps be part of this repo. Assuming they aren't too massive. I also am considering just having the extension repos code as part of this repo, maybe a submodule? That just feels wrong though.